bluetooth: shell: move `bt_shell_private.c` to `common` folder
Relocated `bt_shell_private.c` to the `common` folder to enable
independent use between `BT_SHELL` and `BT_MESH_SHELL`.
