+++ subsys/bluetooth/Kconfig
@@ -236,6 +236,7 @@ rsource "Kconfig.logging"
 config BT_SHELL
 	bool "Bluetooth shell"
 	select SHELL
+	select BT_PRIVATE_SHELL
 	select BT_TICKER_NEXT_SLOT_GET if BT_LL_SW_SPLIT
 	help
 	  Activate shell module that provides Bluetooth commands to the
