+++ subsys/bluetooth/common/bt_shell_private.c
