+++ subsys/bluetooth/host/shell/gatt.c
@@ -27,7 +27,7 @@
 #include <zephyr/shell/shell.h>
 
 #include "host/shell/bt.h"
-#include "host/shell/bt_shell_private.h"
+#include "common/bt_shell_private.h"
 
 #if defined(CONFIG_BT_GATT_CLIENT) || defined(CONFIG_BT_GATT_DYNAMIC_DB)
 extern uint8_t selected_id;
